Fix non-menu transitions between routes of the same orientation
Before, if the user moved between two routes that had the same
orientation while the menu was not open (e.g. with the browser back
button), the route transition would wait forever for the “light panel is
finished animating” callback, which would never run because no animation
would occur.
Now, this edge case is detected proactively, and the transition function
will not wait for animation to complete if the light panel is not going
to move.